feature(PIL-1581): colocalise les configs sujet avec leurs use cases
- indicateurIndividuConfig → indicateur/commands/creerIndicateurIndividuCommentaire.ts
- panierIndividuConfig → panier/commands/creerPanierIndividuCommentaire.ts
- panierConfig → panier/commands/creerPanierCommentaire.ts
- lister* importent la config depuis le creer* correspondant
- commentaire/sujets.ts ne garde que le type SujetCommentaireConfig
