+++ apps/mb-api/src/indicateur/queries/listerIndicateurIndividuCommentaires.ts
@@ -1,7 +1,7 @@
 import { type ListerCommentairesQuery } from '@pilote/mb-shared/commentaire'
 
 import { listerCommentaires } from '@/commentaire/queries/listerCommentaires'
-import { indicateurIndividuConfig } from '@/commentaire/sujets'
+import { indicateurIndividuConfig } from '@/indicateur/commands/creerIndicateurIndividuCommentaire'
 
 type Input = {
   params: { indicateurId: string; individuId: string }
